# openapi.yaml
openapi: 3.0.0
info:
  title: Mitosis API
  description: |
    Mitosis is a unified distributed transport evaluation framework that provides REST API endpoints 
    for managing users, groups, workers, and tasks in a distributed computing environment.

    ## Authentication
    Most endpoints require JWT bearer token authentication. Get a token by calling `/login` with 
    valid credentials.

    ## API Structure
    - `/auth` - User authentication status
    - `/login` - User login
    - `/users` - User management
    - `/groups` - Group management and file attachments
    - `/workers` - Worker registration and management
    - `/tasks` - Task submission and monitoring
    - `/admin` - Administrative operations (admin users only)
    - `/health` - Service health check
    - `/redis` - Redis connection information
  version: 0.6.0
  contact:
    name: Mitosis
    url: https://github.com/stack-rs/mitosis
  license:
    name: Apache 2.0
    url: https://github.com/stack-rs/mitosis/blob/main/LICENSE
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://api.mitosis.example.com
    description: Production server
paths:
...
  /users/groups:
    get:
      summary: Query user groups
      description: Get all groups the authenticated user has access to
      operationId: queryUserGroups
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User groups with roles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupsQueryResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /groups:
    post:
      summary: Create group
      description: Create a new group (user becomes admin of the group)
      operationId: createGroup
      tags:
        - Groups
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupReq"
      responses:
        "200":
          description: Group created successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /groups/{group_name}:
    get:
      summary: Get group information
      description: Get detailed information about a specific group
      operationId: getGroup
      tags:
        - Groups
      security:
        - bearerAuth: []
      parameters:
        - name: group_name
          in: path
          required: true
          schema:
            type: string
          example: research-team
      responses:
        "200":
          description: Group information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupQueryInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /groups/{group_name}/users:
    put:
      summary: Update user group roles
      description: Add users to group or update their roles
      operationId: updateUserGroup
      tags:
        - Groups
      security:
        - bearerAuth: []
      parameters:
        - name: group_name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserGroupRoleReq"
      responses:
        "200":
          description: User roles updated successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Remove users from group (query params)
      description: Remove users from the group using query parameters
      operationId: removeUserGroupParams
      tags:
        - Groups
      security:
        - bearerAuth: []
      parameters:
        - name: group_name
          in: path
          required: true
          schema:
            type: string
        - name: users
          in: query
          required: false
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
          description: Comma-separated list of usernames to remove
      responses:
        "200":
          description: Users removed successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # /groups/{group_name}/users/remove:
  #   post:
  #     summary: Remove users from group (JSON body)
  #     description: Remove users from the group using JSON request body
  #     operationId: removeUserGroup
  #     tags:
  #       - Groups
  #     security:
  #       - bearerAuth: []
  #     parameters:
  #       - name: group_name
  #         in: path
  #         required: true
  #         schema:
  #           type: string
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             $ref: "#/components/schemas/RemoveUserGroupRoleReq"
  #     responses:
  #       "200":
  #         description: Users removed successfully
  #       "401":
  #         $ref: "#/components/responses/Unauthorized"
  #       "403":
  #         $ref: "#/components/responses/Forbidden"
  #       "404":
  #         $ref: "#/components/responses/NotFound"
  /groups/{group_name}/attachments:
    post:
      summary: Upload group attachment
      description: Upload a file to the group's shared storage
      operationId: uploadAttachment
      tags:
        - Groups
      security:
        - bearerAuth: []
      parameters:
        - name: group_name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadAttachmentReq"
      responses:
        "200":
          description: Upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadAttachmentResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /groups/{group_name}/attachments/query:
    post:
      summary: Query group attachments
      description: List attachments in the group with optional filtering
      operationId: queryAttachments
      tags:
        - Groups
      security:
        - bearerAuth: []
      parameters:
        - name: group_name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachmentsQueryReq"
      responses:
        "200":
          description: List of attachments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachmentsQueryResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /groups/{group_name}/attachments/{key}:
    get:
      summary: Get attachment metadata
      description: Get metadata for a specific attachment
      operationId: queryAttachment
      tags:
        - Groups
      security:
        - bearerAuth: []
      parameters:
        - name: group_name
          in: path
          required: true
          schema:
            type: string
        - name: key
          in: path
          required: true
          schema:
            type: string
          style: simple
          explode: false
      responses:
        "200":
          description: Attachment metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachmentMetadata"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Delete attachment
      description: Delete a file from group's shared storage
      operationId: deleteAttachment
      tags:
        - Groups
      security:
        - bearerAuth: []
      parameters:
        - name: group_name
          in: path
          required: true
          schema:
            type: string
        - name: key
          in: path
          required: true
          schema:
            type: string
          style: simple
          explode: false
      responses:
        "200":
          description: Attachment deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /groups/{group_name}/download/attachments/{key}:
    get:
      summary: Download attachment
      description: Get download URL for an attachment
      operationId: downloadAttachment
      tags:
        - Groups
      security:
        - bearerAuth: []
      parameters:
        - name: group_name
          in: path
          required: true
          schema:
            type: string
        - name: key
          in: path
          required: true
          schema:
            type: string
          style: simple
          explode: false
      responses:
        "302":
          description: Redirect to download URL
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /groups/{group_name}/download/attachments:
    post:
      summary: Batch download attachments by filter
      description: Get download URLs for group attachments matching filter criteria
      operationId: batchDownloadAttachmentsByFilter
      tags:
        - Groups
      security:
        - bearerAuth: []
      parameters:
        - name: group_name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachmentsDownloadByFilterReq"
      responses:
        "200":
          description: List of attachment download URLs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachmentsDownloadListResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /groups/{group_name}/download/attachments/list:
    post:
      summary: Batch download attachments by keys
      description: Get download URLs for group attachments by specific keys
      operationId: batchDownloadAttachmentsByKeys
      tags:
        - Groups
      security:
        - bearerAuth: []
      parameters:
        - name: group_name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachmentsDownloadByKeysReq"
      responses:
        "200":
          description: List of attachment download URLs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachmentsDownloadListResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /workers/{uuid}/groups:
    put:
      summary: Update worker group roles
      description: Add groups to worker or update their roles
      operationId: updateWorkerGroups
      tags:
        - Workers
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroupWorkerRoleReq"
      responses:
        "200":
          description: Group roles updated successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Remove worker group roles (query params)
      description: Remove groups from worker using query parameters
      operationId: removeWorkerGroupsParams
      tags:
        - Workers
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: groups
          in: query
          required: false
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
          description: Comma-separated list of group names to remove
      responses:
        "200":
          description: Groups removed successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # /workers/{uuid}/groups/remove:
  #   post:
  #     summary: Remove worker group roles (JSON body)
  #     description: Remove groups from worker using JSON request body
  #     operationId: removeWorkerGroups
  #     tags:
  #       - Workers
  #     security:
  #       - bearerAuth: []
  #     parameters:
  #       - name: uuid
  #         in: path
  #         required: true
  #         schema:
  #           type: string
  #           format: uuid
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             $ref: "#/components/schemas/RemoveGroupWorkerRoleReq"
  #     responses:
  #       "200":
  #         description: Groups removed successfully
  #       "401":
  #         $ref: "#/components/responses/Unauthorized"
  #       "403":
  #         $ref: "#/components/responses/Forbidden"
  #       "404":
  #         $ref: "#/components/responses/NotFound"
...
  /admin/groups/{group_name}/attachments/{key}:
    delete:
      summary: Delete group attachment (Admin)
      description: Delete any group attachment (admin only)
      operationId: adminDeleteAttachment
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: group_name
          in: path
          required: true
          schema:
            type: string
        - name: key
          in: path
          required: true
          schema:
            type: string
          style: simple
          explode: false
      responses:
        "200":
          description: Attachment deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /login endpoint
    workerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained when registering worker
  responses:
...
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid or expired token
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Insufficient permissions
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Resource not found
    Conflict:
      description: Conflict - resource already exists or state conflict
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Resource already exists
  schemas:
...
    GroupState:
      type: string
      enum:
        - Active
        - Locked
        - Deleted
      description: |
        Group state:
        - Active: Group is active and can be used
        - Locked: Group is temporarily disabled
        - Deleted: Group is marked for deletion
...
    UserGroupRole:
      type: string
      enum:
        - Read
        - Write
        - Admin
      description: |
        User's role in a group:
        - Read: Can view group information and tasks
        - Write: Can submit tasks and create workers for the group
        - Admin: Can manage group membership and settings
...
    AttachmentContentType:
      type: string
      enum:
        - application/octet-stream
      description: MIME type of group attachment
...
    CreateGroupReq:
      type: object
      required:
        - group_name
      properties:
        group_name:
          type: string
          description: Name of the new group
          example: research-team
    GroupQueryInfo:
      type: object
      required:
        - group_name
        - creator_username
        - created_at
        - updated_at
        - state
        - task_count
        - storage_quota
        - storage_used
        - worker_count
      properties:
        group_name:
          type: string
          example: research-team
        creator_username:
          type: string
          example: johndoe
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        state:
          $ref: "#/components/schemas/GroupState"
        task_count:
          type: integer
          format: int64
          description: Number of tasks in the group
          example: 42
        storage_quota:
          type: integer
          format: int64
          description: Storage quota in bytes
          example: 1073741824
        storage_used:
          type: integer
          format: int64
          description: Storage currently used in bytes
          example: 536870912
        worker_count:
          type: integer
          format: int64
          description: Number of workers serving this group
          example: 5
        users_in_group:
          type: object
          nullable: true
          additionalProperties:
            $ref: "#/components/schemas/UserGroupRole"
          description: Users in the group and their roles
          example:
            johndoe: Admin
            janedoe: Write
    GroupsQueryResp:
      type: object
      required:
        - groups
      properties:
        groups:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/UserGroupRole"
          description: Groups the user has access to and their role
          example:
            my-group: Admin
            research-team: Write
    UpdateUserGroupRoleReq:
      type: object
      required:
        - relations
      properties:
        relations:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/UserGroupRole"
          description: Username to role mapping
          example:
            johndoe: Admin
            janedoe: Write

    # RemoveUserGroupRoleReq:
    #   type: object
    #   required:
    #     - users
    #   properties:
    #     users:
    #       type: array
    #       items:
    #         type: string
    #       description: Usernames to remove from group
    #       example:
    #         - johndoe
    #         - janedoe
...
    UploadAttachmentReq:
      type: object
      required:
        - key
        - content_length
      properties:
        key:
          type: string
          description: Unique key for the attachment
          example: datasets/experiment-1.tar.gz
        content_length:
          type: integer
          format: uint64
          description: Size of file in bytes
          example: 1048576
    UploadAttachmentResp:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          description: Presigned URL for uploading the file
          example: https://s3.example.com/bucket/path?signature=...
    AttachmentMetadata:
      type: object
      required:
        - content_type
        - size
        - created_at
        - updated_at
      properties:
        content_type:
          $ref: "#/components/schemas/AttachmentContentType"
        size:
          type: integer
          format: int64
          description: File size in bytes
          example: 1048576
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    AttachmentsQueryReq:
      type: object
      properties:
        key:
          type: string
          nullable: true
          description: Filter attachments by part of the key
          example: datasets/
        limit:
          type: integer
          format: uint64
          nullable: true
          description: Maximum number of results
          example: 50
        offset:
          type: integer
          format: uint64
          nullable: true
          description: Offset for pagination
          example: 0
        count:
          type: boolean
          description: Whether to return total count
          default: false
    AttachmentQueryInfo:
      type: object
      required:
        - key
        - content_type
        - size
        - created_at
        - updated_at
      properties:
        key:
          type: string
          example: datasets/experiment-1.tar.gz
        content_type:
          $ref: "#/components/schemas/AttachmentContentType"
        size:
          type: integer
          format: int64
          example: 1048576
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    AttachmentsQueryResp:
      type: object
      required:
        - count
        - attachments
        - group_name
      properties:
        count:
          type: integer
          format: uint64
          description: Total number of attachments (if count=true)
          example: 25
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/AttachmentQueryInfo"
        group_name:
          type: string
          example: research-team
...
    UpdateGroupWorkerRoleReq:
      type: object
      required:
        - relations
      properties:
        relations:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/GroupWorkerRole"
          description: Group to role mapping
          example:
            research-team: Write
            ml-group: Admin

    # RemoveGroupWorkerRoleReq:
    #   type: object
    #   required:
    #     - groups
    #   properties:
    #     groups:
    #       type: array
    #       items:
    #         type: string
    #       description: Groups to remove from worker
    #       example:
    #         - old-group
    #         - deprecated-group
...
    AttachmentsDownloadByFilterReq:
      type: object
      properties:
        key:
          type: string
          nullable: true
          description: Filter attachments by key pattern
          example: datasets/
        limit:
          type: integer
          format: uint64
          nullable: true
          description: Maximum number of results
          example: 50
        offset:
          type: integer
          format: uint64
          nullable: true
          description: Offset for pagination
          example: 0
      description: Request to batch download attachments by filter criteria
    AttachmentsDownloadByKeysReq:
      type: object
      required:
        - keys
      properties:
        keys:
          type: array
          items:
            type: string
          description: List of attachment keys to download
          example:
            - datasets/experiment-1.tar.gz
            - models/checkpoint.pth
      description: Request to batch download attachments by specific keys
    AttachmentDownloadItem:
      type: object
      required:
        - key
        - url
        - size
      properties:
        key:
          type: string
          description: Attachment key
          example: datasets/experiment-1.tar.gz
        url:
          type: string
          format: uri
          description: Presigned download URL
          example: https://s3.example.com/bucket/path?signature=...
        size:
          type: integer
          format: int64
          description: Attachment size in bytes
          example: 1048576
    AttachmentsDownloadListResp:
      type: object
      required:
        - downloads
        - group_name
      properties:
        downloads:
          type: array
          items:
            $ref: "#/components/schemas/AttachmentDownloadItem"
          description: List of attachment downloads
        group_name:
          type: string
          description: Group name
          example: research-team
security:
  - bearerAuth: []
...